All crypto tests passing and working on fpga
