+++ software/src/crypto/McEliece/pk_gen.c
@@ -24,155 +24,6 @@
 #include "printf.h"
 #include "arena.h"
 
-static VectorLikeOperationConfig* vec;
-static void* matAddr;
-
-// Matrix Allocate
-#define MA(Y,X,TYPE) (TYPE*) PushBytes(sizeof(TYPE) * (X) * (Y))
-
-// Matrix Index
-#define MI(Y,X,ROWSIZE) ((Y) * (ROWSIZE) + (X)) 
-#define MR(Y,ROWSIZE) ((Y) * (ROWSIZE))
-
-#define ALIGN_4(VAL) (((VAL) + 3) & (~3))
-#define ALIGN_16(VAL) (((VAL) + 15) & (~15))
-
-// Need to align
-#define ASYS (SYS_N / 8) // 436 -> 448
-#define SBYTE (SYS_N / 8) // 436 -> 448
-
-//#define ASYS ALIGN_16(SYS_N / 8) // 436 -> 448
-//#define SBYTE ALIGN_4(SYS_N / 8) // 436 -> 448
-#define SINT (ASYS / 4) // 109 -> 112
-
-void PrintRow(unsigned char* view){
-#if 0
-    for(int i = 0; i < SBYTE; i++){
-        printf("%02x",view[i]);
-        if(i % 16 == 0 && i != 0){
-            printf("\n");
-        }
-    }
-#else
-    for(int i = 0; i < 16; i++){
-        printf("%02x",view[i]);
-    }
-    printf("\n");
-    for(int i = SBYTE - 16; i < SBYTE; i++){
-        printf("%02x",view[i]);
-    }
-    printf("\n");
-#endif
-}
-
-void PrintFullRow(unsigned char* view){
-    for(int i = 0; i < SBYTE; i++){
-        printf("%02x",view[i]);
-        if(i % 16 == 0 && i != 0){
-            printf("\n");
-        }
-    }
-}
-
-void ReadRow(uint32_t* row){
-    for (int i = 0; i < SINT; i++){
-        row[i] = VersatUnitRead((iptr) matAddr,i);
-    }
-}
-
-void VersatLoadRow(uint32_t* row){
-    VersatMemoryCopy(matAddr,(int*) row,SINT * sizeof(int));
-}
-
-void VersatPrintRow(){
-    uint32_t values[SINT];
-    ReadRow(values);
-    PrintRow((uint8_t*) values);
-}
-
-uint8_t* GetVersatRowForTests(){  
-    static uint8_t buffer[ASYS];
-    ReadRow((uint32_t*) buffer);
-    return buffer;
-}
-
-void PrintSimpleMat(unsigned char** mat,int centerRow){
-    PrintRow(mat[0]);
-    PrintRow(mat[centerRow - 1]);
-    PrintRow(mat[centerRow]);
-    PrintRow(mat[centerRow + 1]);
-    PrintRow(mat[PK_NROWS - 1]);
-}
-
-void PrintFullMat(unsigned char** mat){
-    printf("Printing full mat:\n");
-    for(int i = 0; i < PK_NROWS; i++){
-        PrintFullRow(mat[i]);
-    }
-}
-
-void VersatMcElieceLoop1(uint8_t *row, uint8_t mask,bool first){
-    static uint8_t savedMask = 0;
-    uint32_t *row_int = (uint32_t*) row;
-
-    ConfigureSimpleVReadShallow(&vec->row, SINT, (int*) row_int);
-    if(first){
-        vec->mat.in0_wr = 0;
-    } else {
-        uint32_t mask_int = (savedMask) | (savedMask << 8) | (savedMask << 8*2) | (savedMask << 8*3);
-        vec->mask.constant = mask_int;
-        vec->mat.in0_wr = 1;
-    }
-
-    EndAccelerator();
-    StartAccelerator();
-
-    savedMask = mask;
-}
-
-// Can only be called if k != row. Care
-void VersatMcElieceLoop2(unsigned char** mat,int timesCalled,int k,int row,uint8_t mask){
-    static uint8_t savedMask = 0;
-
-    int toRead =    k;
-    int toCompute = ((toRead - 1    == row) ? toRead - 2    : toRead - 1);
-    int toWrite =   ((toCompute - 1 == row) ? toCompute - 2 : toCompute - 1);
-
-    if(toRead < PK_NROWS){
-        int *toRead_int = (int*) mat[toRead];
-
-        vec->mat.in0_wr = 0;
-
-        ConfigureSimpleVReadShallow(&vec->row, SINT,toRead_int);        
-    } else {
-        vec->row.enableRead = 0;
-        toRead = -9;
-    }
-    
-    if(timesCalled >= 1 && toCompute >= 0 && toCompute < PK_NROWS){
-        uint32_t mask_int = (savedMask) | (savedMask << 8) | (savedMask << 8*2) | (savedMask << 8*3);
-
-        vec->mask.constant = mask_int;
-    } else {
-        ConfigureSimpleVWrite(&vec->writer, SINT, (int*) NULL);
-        vec->writer.enableWrite = 0;
-        toCompute = -9;
-    }
-    
-    if(timesCalled >= 2 && toWrite >= 0){
-        int* toWrite_int = (int*) mat[toWrite];
-        ConfigureSimpleVWrite(&vec->writer, SINT,toWrite_int);
-    } else {
-        toWrite = -9;
-        vec->writer.enableWrite = 0;
-    }
-
-    EndAccelerator();
-    StartAccelerator();
-
-    savedMask = mask;
-}
-
 static crypto_uint64 uint64_is_equal_declassify(uint64_t t, uint64_t u) {
     crypto_uint64 mask = crypto_uint64_equal_mask(t, u);
     crypto_declassify(&mask, sizeof mask);
@@ -193,28 +44,11 @@ int pk_gen(unsigned char *pk, unsigned char *sk, const uint32_t *perm, int16_t *
 
     int mark = MarkArena();
 
-    // Init needed values for versat later on.  
-    vec = (VectorLikeOperationConfig*) accelConfig;
-    matAddr = TOP_mat_addr;
-
-    ActivateMergedAccelerator(VectorLikeOperation);
-
-    ConfigureSimpleVReadBare(&vec->row);
-
-    vec->mat.iterA = 1;
-    vec->mat.incrA = 1;
-    vec->mat.iterB = 1;
-    vec->mat.incrB = 1;
-    vec->mat.perA = SINT + 1;
-    vec->mat.dutyA = SINT + 1;
-    vec->mat.perB = SINT + 1;
-    vec->mat.dutyB = SINT + 1;
-
     uint64_t buf[ 1 << GFBITS ];
 
-    volatile unsigned char** mat = PushArray(PK_NROWS,volatile unsigned char*);
+    unsigned char** mat = PushArray(PK_NROWS,unsigned char*);
     for(int i = 0; i < PK_NROWS; i++){
-        mat[i] = PushArray(SYS_N / 8,volatile unsigned char);
+        mat[i] = PushArray(SYS_N / 8,unsigned char);
     }
 
     unsigned char mask;
@@ -306,31 +140,6 @@ int pk_gen(unsigned char *pk, unsigned char *sk, const uint32_t *perm, int16_t *
                 break;
             }
 
-#if 1
-            uint32_t *out_int = (uint32_t*) mat[row];
-            EndAccelerator();
-
-            VersatLoadRow(out_int);
-            bool first = true;
-            for (k = row + 1; k < PK_NROWS; k++) {
-                mask = mat[ row ][ i ] ^ mat[ k ][ i ];
-                mask >>= j;
-                mask &= 1;
-                mask = -mask;
-
-                VersatMcElieceLoop1(mat[k],mask,first);
-
-                // We could fetch the value from Versat, but it's easier to calculate it CPU side.
-                mat[row][i] ^= mat[k][i] & mask;
-                first = false;
-            }
-
-            // Last run, use valid data to compute last operation
-            VersatMcElieceLoop1(mat[PK_NROWS - 1],0,false); // TODO: Have a proper function instead of sending a "fake" adress
-
-            EndAccelerator();
-
-#else
             for (k = row + 1; k < PK_NROWS; k++) {
                 mask = mat[ row ][ i ] ^ mat[ k ][ i ];
                 mask >>= j;
@@ -341,34 +150,13 @@ int pk_gen(unsigned char *pk, unsigned char *sk, const uint32_t *perm, int16_t *
                     mat[ row ][ c ] ^= mat[ k ][ c ] & mask;
                 }
             }
-#endif
+
             if ( uint64_is_zero_declassify((mat[ row ][ i ] >> j) & 1) ) { // return if not systematic
                 printf("Early finish row:%d\n",row);
                 PopArena(mark);
                 return -1;
             }
 
-            ReadRow(out_int);
-
-#if 1
-            int index = 0;
-            for (k = 0; k < PK_NROWS; k++) {
-                if (k != row) {
-                    mask = mat[k][i] >> j;
-                    mask &= 1;
-                    mask = -mask;
-
-                    VersatMcElieceLoop2(mat,index,k,row,mask);
-                    index += 1;
-                }
-            }
-
-            VersatMcElieceLoop2(mat,index++,PK_NROWS,row,0);
-            VersatMcElieceLoop2(mat,index++,PK_NROWS + 1,row,0);
-            vec->writer.enableWrite = 0;
-
-            clear_cache();
-#else
             for (k = 0; k < PK_NROWS; k++) {
                 if (k != row) {
                     mask = mat[ k ][ i ] >> j;
@@ -380,7 +168,6 @@ int pk_gen(unsigned char *pk, unsigned char *sk, const uint32_t *perm, int16_t *
                     }
                 }
             }
-#endif
         }
     }
