+++ software/sw_build.mk
@@ -80,8 +80,8 @@ IOB_SOC_OPENCRYPTOLINUX_LFLAGS=-Wl,-Bstatic,-T,$(TEMPLATE_LDS),--strip-debug
 IOB_SOC_OPENCRYPTOLINUX_FW_SRC=src/iob_soc_opencryptolinux_firmware.S
 IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=src/iob_soc_opencryptolinux_firmware.c
 IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=src/printf.c
-IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=src/SHA_AES.c
-#IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=src/crypto/sha2.c
+IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=src/versat_crypto.c
+IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=src/versat_crypto_tests.c
 IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=src/crypto/aes.c
 
 IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=$(wildcard src/crypto/McEliece/*.c)
